requirements.txt
numpy==1.23.4
scipy==1.9.3
pandas==1.3.5
protobuf==3.20.3
python-dotenv
streamlit==1.26.0
google-auth==1.35.0
openai>=0.28.1
langid==1.1.6
typing-inspect==0.8.0
typing_extensions>=4.5.0
pydantic==1.10.11
langchain==0.0.200
selenium
seleniumbase
webdriver-manager
streamlit-toggle-switch
# hotel pdf parser
pytesseract
pdf2image
stqdm==0.0.5
selenium
seleniumbase
webdriver-manager
apify==1.1.4
requests
serpapi
pinecone
pinecone-text